+++ CHANGES.md
@@ -17,6 +17,20 @@ To be released.
 [#3]: https://github.com/dahlia/upyo/issues/3
 
 
+Version 0.1.1
+-------------
+
+Released on July 14, 2025.
+
+### @upyo/smtp
+
+ -  Fixed CJK character encoding corruption in SMTP transport HTML emails.
+    Korean, Japanese, and Chinese characters are now properly encoded using
+    UTF-8 quoted-printable encoding.  [[#4]]
+
+[#4]: https://github.com/dahlia/upyo/issues/4
+
+
 Version 0.1.0
 -------------